separate run and default tasks
